LUCENE-9302,SOLR-14381: Integer overflow in total count in grouping r… (#29)
* LUCENE-9302,SOLR-14381: Integer overflow in total count in grouping results

* updated lucene version

* updated lucene branch and interface

* updated

* spotless apply

* updated test

+++ solr/core/src/java/org/apache/solr/search/grouping/CommandHandler.java
@@ -123,7 +123,7 @@ public class CommandHandler {
   private final boolean truncateGroups;
   private final boolean includeHitCount;
   private boolean partialResults = false;
-  private int totalHitCount;
+  private long totalHitCount;
 
   private DocSet docSet;
 
@@ -266,7 +266,7 @@ public class CommandHandler {
     }
   }
 
-  public int getTotalHitCount() {
+  public long getTotalHitCount() {
     return totalHitCount;
   }
 }
