+++ solr/core/src/java/org/apache/solr/search/grouping/distributed/command/SearchGroupsFieldCommandResult.java
@@ -23,16 +23,16 @@ import org.apache.lucene.util.BytesRef;
 /** Encapsulates the result of a {@link SearchGroupsFieldCommand} command */
 public class SearchGroupsFieldCommandResult {
 
-  private final Integer groupCount;
+  private final Long groupCount;
   private final Collection<SearchGroup<BytesRef>> searchGroups;
 
   public SearchGroupsFieldCommandResult(
-      Integer groupCount, Collection<SearchGroup<BytesRef>> searchGroups) {
+      Long groupCount, Collection<SearchGroup<BytesRef>> searchGroups) {
     this.groupCount = groupCount;
     this.searchGroups = searchGroups;
   }
 
-  public Integer getGroupCount() {
+  public Long getGroupCount() {
     return groupCount;
   }
